feat : merge pdf and docx in single endpoint
